+++ Assets/Audio/AudioObjects/Jokes/allJokes.asset
@@ -90,4 +90,4 @@ MonoBehaviour:
     diffusion: 0
     density: 0
   neverRepeat: 1
-  lastClipIndex: 3
+  lastClipIndex: 4
