+++ Assets/Prefabs/Menu/MenuParent.prefab
@@ -320,11 +320,11 @@ MonoBehaviour:
       - m_Target: {fileID: 6632191372605382610}
         m_TargetAssemblyTypeName: MenuSystem2, Assembly-CSharp
         m_MethodName: SceneSwitch
-        m_Mode: 5
+        m_Mode: 3
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
           m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-          m_IntArgument: 0
+          m_IntArgument: 1
           m_FloatArgument: 0
           m_StringArgument: StartLevel
           m_BoolArgument: 0
