+++ Assets/Scenes/Level_1.unity
@@ -63869,7 +63869,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 3900558465337787480, guid: b30eb0ec257ff674a8a7936b3faaa51d, type: 3}
       propertyPath: fadeBehaviour
-      value: 1
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 7749441706675402418, guid: b30eb0ec257ff674a8a7936b3faaa51d, type: 3}
       propertyPath: listener
