--- Assets/Scripts/MenuSystem2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a74d3df537869784ebc97eb799c1b976
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
