+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,10 +6,10 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/Scenes/Menu.unity
+    path: Assets/Scenes/StartMenu.unity
     guid: 6dc192d0ef33a5a4fb6955f35add6e32
   - enabled: 1
-    path: Assets/Scenes/StartLevel.unity
+    path: Assets/Scenes/Level_1.unity
     guid: c4eee1fc2e26dbe4b9195ae25fabe621
   m_configObjects:
     com.unity.input.settings: {fileID: 11400000, guid: 9e7be553448fa2546aea5752021cbcf7, type: 2}
